+++ dashboard/assets/js/app.js
@@ -7,10 +7,17 @@ import {LiveSocket} from "phoenix_live_view"
 
 let csrfToken = document.querySelector("meta[name='csrf-token']").getAttribute("content")
 
+let Hooks = {}
+
+Hooks.ScrollBottom = {
+  mounted() { this.el.scrollTop = this.el.scrollHeight },
+  updated() { this.el.scrollTop = this.el.scrollHeight }
+}
+
 let liveSocket = new LiveSocket("/live", Socket, {
   longPollFallbackMs: 2500,
   params: {_csrf_token: csrfToken},
-  hooks: {}
+  hooks: Hooks
 })
 
 liveSocket.connect()